requirements.txt
openai==1.12.0
python-dotenv==1.0.0
flask==2.3.3
pydantic==2.5.0
bleach==6.1.0
ratelimit==2.2.1
psutil==5.9.0